Adds ability to match a device according to different rules
A class that implements IUsbDeviceMatchable has to be used in the constructor of the UsbHidDevice instead of the previous constraing VID, PID. This allows the device to be matched against differnt types of rules.

Two matcher classes are include, one that matches against the VID/PID like before and one that matches the serial string.

Still only the first device to match the rules will be returned (no multiple device support yet).